feat: implement core system architecture and admin/driver/rider management
This major commit introduces the foundational architecture and user management systems for the platform:

1. Admin Management System:
- Added Admin model, controller, and auth system
- Implemented admin-specific commands (CreateAdminAccount)
- Added admin notification system (two-factor, password reset)

2. User Authentication:
- Created separate auth systems for Admin/Driver/Rider
- Implemented Two-Factor Authentication for all user types
- Added profile management controllers

3. Core Services & Resources:
- Implemented service layer for all major domains (Car, Trip, Payment, etc.)
- Created API Resources for consistent JSON responses
- Added request validation classes for all endpoints

4. Database Structure:
- Restructured database migrations and seeders
- Removed old Car model in favor of DriverCar approach
- Added proper relationships between models

5. Localization:
- Added Arabic/English language files
- Implemented translation helpers

6. Utility Features:
- Added system-wide traits (Suspendable, TwoFactorCodeGenerator)
- Implemented notification services
- Added custom validation rules

BREAKING CHANGES:
- Removed old AuthTemplate controllers
- Deleted legacy Car model and related files
- Restructured database factories and seeders
- Modified existing migrations for new structure

+++ database/migrations/domains/vehicles/pricings/2025_06_08_10004_create_pricings_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
 
             $table->boolean('is_active')->default(true);
 
-            $table->decimal('price_per_km');
+            $table->integer('price_per_km');
 
             $table->timestamps();
         });
